src/screens/components/dialog_attendance_result.py
import streamlit as st
import pandas as pd
from src.screens.database.db import create_subject
from src.screens.database.config import supabase
from src.screens.database.db import enroll_student_to_subject
import time
from PIL import Image
from src.screens.database.db import create_attendance


def show_attendance_result(df,logs):  
    st.write('Please review the attendance results before confirming.')
    st.dataframe(df,hide_index=True,width='stretch')

    col1,col2=st.columns(2)

    with col1:
        if st.button('Discard', width='stretch'):
            st.session_state.attendance_images=[] 
            st.session_state.voice_attendance_results=None
            st.rerun()

    with col2:
        if st.button('Confirm & save', width='stretch',type='primary'):
            try:
                create_attendance(logs)
                st.toast('Attendance saved successfully')
                st.session_state.attendance_images=[] 
                st.session_state.voice_attendance_results=None
                st.rerun()
                
            except Exception as e:
                st.error(f'Error saving attendance: ')
                st.rerun()


@st.dialog("Attendance Result")
def attendance_result_dialog(df,logs):
    show_attendance_result(df,logs)