requirements.txt
streamlit

#utilities

numpy
pandas
#face recognition
scikit-learn 
dlib-bin
git+https://github.com/ageitgey/face_recognition_models

setuptools

supabase 
bcrypt
segno
pillow
streamlit

libross
resemblyzer
